[4-auth-usuarios-admin] chore: Rename get-session to auth-server
